Refactor game state management: Split store into slices for hot simulation, progression, and UI
- Created separate slices for hot simulation, progression, and UI state management in the game store.
- Updated the game store to utilize these slices, improving modularity and maintainability.
- Enhanced code coverage metrics in the coverage reports.
- Updated documentation to reflect changes in state management structure.
